+++ test/testdata/render_options.py
@@ -1,0 +1,1 @@
+"""Testing some render options. For example, there is no 'View Source' button here."""