+++ pdoc/__init__.py
@@ -346,7 +346,7 @@ We would be happy to accept contributions for Markdown and PDF.
 '''
 from __future__ import annotations
 
-__version__ = "7.2.0"  # this is read from setup.py
+__version__ = "7.3.0"  # this is read from setup.py
 
 import io
 import traceback