+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 # Unreleased: pdoc next
 
- - 
+ - Full compatibility with Python 3.10.
+ - Many of pdoc's template customizations are now available directly as command line switches,
+   for example `--math`.
 
 # 2021-07-28: pdoc 7.2.0